backend/pyproject.toml
[project]
name = "ask-docs-backend"
version = "0.1.0"
description = "RAG-based Q&A backend with local LLM"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sse-starlette>=2.0.0",
    "httpx>=0.26.0",
    # Document parsing
    "pymupdf>=1.23.0",
    "beautifulsoup4>=4.12.0",
    "openpyxl>=3.1.0",
    "markdown>=3.5.0",
    "lxml>=5.0.0",
    # ML/Embeddings
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "torch>=2.1.0",
    "numpy>=1.26.0",
    # Reranking
    "transformers>=4.36.0",
    # LLM
    "ollama>=0.1.0",
    "llama-cpp-python>=0.2.0",
    # Observability
    "structlog>=23.2.0",
    "python-json-logger>=2.0.0",
    # Caching & Rate limiting
    "cachetools>=5.3.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
